Decided that narrative stuff should really be allowed at the section level too.  For example, a LaTeX file typically starts with a prelude and an introduction, before the first Z section starts.
This required adding a new kind of section, which made three kinds
of sections (normal, unparsed and narrative).  Obviously a common
supertype was needed for these, so we now have <Section> as
the abstract element, with <ZSection> as a concrete subtype that
contains parsed Z stuff, <UnparsedZSection> which contains totally
unparsed Z stuff, and <NarrativeSection> which contains
text/html/unicode/LaTeX etc.  (It might be nice to add an attribute
somewhere to specify what kind of content is inside the narrative section).