[This message is for the previous commit: v1.10 -> v1.11]
Changed the formal parameters of a class to be a list of DeclName,
rather than a list of RefName, because they do declare new names.
This makes formal parameters of classes consistent with the formal parameters
of schemas in standard Z, as pointed out by Tim Miller.

Also changed the defn of RenameOpExpr to add a list of name/name
pairs rather than a list of RenameExpr.  This removes the unnecessary
extra Expr and makes it more similar to the standard Z defn of RenameExpr.