BUGFIX: rename exprs now seem to work correctly for Z, but this has introduced a problem in OZ operation expressions.
Removed all references to WHERE and replaced with BAR. Remove "\\where"
from the scanner.

Symbol names declared in the parser now match the names in Petra's
unicode scanner. This is to make the transition to the unicode scanner a
bit smoother :) The only difference seems to be STROKE, for which I am
using INSTROKE, OUTSTROKE etc.