ADD: Final Circus token terminals ADD:  Comment on the changes from Circus TOKENS to reuse of Z tokens due to unavailable unicode. That is namely, CIRCGUARD, REPSEQ, and CIRCIF. ADD: Process and action calls. ADD: Circus qualified parameters and parameterised and indexed process and parameterised actions ADD: Adjusted precedences for the removed tokens and used on their productions the a corresponding _XXXX token precedence instead of the Z precedence (i.e. used a %prec _XXXX ADD: Parenthised processes and actions ADD: Comments on the choice/decision making for hard space used on communication as in "c~?x" rather than "c?x". REMOVE: Productions for qualified parameters as commands.
OBS: Removed the SchExpr rule as an action. The consequence is that one can only have RefExpr on action (like a call), rather than on-the-fly schemas. This is in the spirit of Spivey's Z rather than the Z Std. The choice is due to an enourmous amount of conflicts (and no guessed solution) for the problem.

OBS: Process renaming still with conflict, as well as coercion/assumption commands.

OBS: Tried restricting the expression subtree according to the rules defined in channels.tex file without luck. So, for the moment, just left the whole expression tree, even though not everything should be allowed/used. I've left the productions commented in the source code. But nobody should create "lambda" or "forall" CSExpr anyway!

OBS: All inclusions incurr in NO conflicts!

OBS: All Circus productions are there, but those which would incur conflict were left commented for the moment. They start with a "//CONFLICTS"  comment line.