MODIFY: Apply Ian's changes as described below ian's email: s/.uk/.nz/ because York is not in New Zealand (unfortunately!).
Keywords: deleted Z, because Z is already in paper's title.

Abstract: resurrected mention of ZLive, since section 6 was resurrected.

3.1 para 1, inserted new third sentence about syntax of provisos.

3.1, para 1, last sentence: "Their informal semantics is that the conclusion
is true if all the antecedents are true" deleted, because it is inappropriate
in the syntax section, and the point is said better later in 3.3.

3.1 LaTeX mark-up, inserted \proviso.

3.1 examples of jokers: capitalised e,f,y to E,F,Y as used earlier.

3.1 last sentence, rephrased so that "forever" changed to throughout
its global scope.

Inserted new section 3.2 about type system.

Renamed 3.2 and 3.3 (as were) as Proviso Semantics and Rule Semantics.

3.2 (as was) last sentence, deleted, because syntax of provisos
is now discussed in 3.1.

3.3 (as was) second paragraph, second sentence: "All provisos must
satisfy the following property." deleted because it was redundant.

4.1 Provisos: moved this subsection to section 3, appended to 3.2 as was,
because it isn't about a prover or rewrite engine.

4.1 (as was) provisos: rephrased to say when they pass,
that known instantiations of jokers are taken into account,
and that unification produces new instantiations of jokers.

4.1 (as was) check proviso: inserted "free".