BUGFIX: bug [ 1910815 ] distinct bindings are treated as equal. This was a major design flaw in FlatOr.  The left env was used for the output env, which meant that the correct values were returned for the left side of the OR, and the correct number of solutions for the right side but with the old values from the left side.
Now FlatOr generates a fresh output env containing just the additional
output variables (which must be the same on both sides).  After each 
solution is found, the values of those output variables are copied
from the local (left or right) env to the output env.

Also added a test that checks the expected values, not just the number of solutions!