UPDATE: OpTable and ThmTable visitor / command / services to take genercs into account (and to understand the protocol), and to document them BUG: basicsm.tex example had the wrong section name
ADD: general testing framework base class. it reproduces protocols from both circus and z typecheckers. it is meant to be a base test class for jUnit
that searches a list of resources, each of which will become either a positive or negative test case, depending on the filename conventions taken.