UPDATE: AbstractDCTerm raises domain check exception - and other places as well UPDATE: DCTerm removes a few parameters - they are all collected in DomainChecker UPDATE: DomainCheckException is now a command exception, inline with parsing/typechecking, etc/ UPDATE: DomainCheckPropKeys added a few new options and normalised other options UPDATE: dc_toolkit.tex with proper LATEX/UNICODE correspondence for appliesTo definitions UPDATE: DomainCheckerTest to make sure the sect manager's Source info is up to date ADD: test case for basic sec managed file UPDATE: ZSectDCEnvAnn (and Spec) is simplified and no longer mentions SEctionManager BUGFIX: path separator from ";" to File.pathSeparator and similarly for default "_dc" extension
UPDATE: Most updates are in DomainChecker and DomainCheckUtils. They involve separation of concerns
and simplification. DomainCheckUtils is top-level functionality for both UI and programable calls. This includes
printing results as a file. DomainChecker is solely responsible for Section Management updates, and proper
collectiono of DC VCs via DCTerm. 

OBS:
Major update on domain checking in preparation for a general Z VCG. This includes great
simplification of the DC protocols, mainly that is does not process whole specifications,
but rather ZSects only. It's up to the user to process multiple ZSect specs.