UPDATE: AbstractVC generates type-correct names (e.g., no $) UPDATE: various refactorings in preparation for feasibility package ADD: distinct DefinitionTable and service ADD: various DC examples and some DefTbl examples
PS:
In time, this definition table could be the one at Parser. At the moment,
it isn't since it does not cater well for jokers and Rules started to raise errors.

Although Definition has a Type, it also works in case of untyped specs (e.g., it is extra info).