ADD: Z/Eves proofs tokenisation / lexing and beginnings of parsing of such tokens. There are a few issues/design decisions yet to be resolved:
a) Z/Eves proof commands/scripts are written in a verbatim-like environment in LaTeX
b) they contain "keywords" that are mixfix and with Z-LaTeX in between them (e.g., "apply" THM-NAME "to expression" Z-EXPR)
c) there is no "Unicode" equivalent to the Z-LaTeX in proofs, although Z/Eves does export proof scripts to Unicode
d) keywords in proof scripts are not "reserved" as such: the user can have ZNames of them, which is unusual
e) there are two categories of proof commands: those related to proof scripts and those related to driving a proving session

Many of these issues are resolved, yet some create conflicting design decisions. For instance, it's harder to use CZT to parse these
keywords, because of the seemingless meaning of both markups, one of which is non-existent in the Z/Eves sources. For now, I am
having a rather similar lexing on both Latex2Unicode and ContextFreeScanner. The fact that keywords are "split" into various words
also creates some lexing complications which I am trying to resolve with a separate LaTeX symbol (PROOFWORD) to be used by the
parser. The KeywordScanner is yet to be finished, given there is a pending decision on whether to have part of proof words as keywords
or not. The other scanners were just cosmetically updated in the usual way for a first commit. Work on Parser.xml will follow soon.

TODO: I am currently working on the Theorem environment, following recent discussions CZT-devel