UPDATE: zevesopt.tex test for the whole set of possible situations ADD: support for ignoring all kinds of operators ADD: do not allow undeclared names in SchText DeclParts even within ConjPara. Only on the QntPred(_, SchText.Pred, Pred) of a ConjPara ADD: various other tests, including the only case of operator template that it's to be ignored (e.g., ingen/pregen types) ADD: nicer status messages during tests + print hidden warnings for reference
ADD: context information to make sure only the right places the error is ignored. That is, only when within ConjPara errors are to be considered to be ignored.
And even within them, in certain cases, they *must not* be ignored, like within the ZDeclPart of QntPred(SchText(ZDeclPart, Pred), Pred) - both Preds can contain
undeclared names.