COSMETICS: clean up scanners ADD: treatment for implicitly defined operator "applies$to". Tricky because of "$" and because "to" is a reserved word
COMMENT: on balance, after trying various options, the best solution is to have applies$to as a toolkit (zeves_toolkit), where the operator name is implicitly declared.
I tried variations where it was ignored at places like Pred and let ZEves handle it but this wasn't satisfactory because it could crop up in other places. The rationale was
to add special treatment at low-level scanners in order to keep the parsing simple.