ADD: Implemented listener support to ZEvesSnapshot - events are fired when results are added, removed and cleared within the snapshot.
ADD: Implemented Z/Eves Tactic support. Currently it replicates functionality of "prove" and "prove by reduce" commands, so that we could get the trace of applied commands. The implementation constructs a command sequence and loops it tracking the results.

ADD: Extended ZEvesOutputView to display result traces. Also updated the partitioning to display the results and traces nicely with appropriate syntax highlighting.