ADD: support for complex schema calculus involving multi-layered renaming/hiding (e.g., this lends itself to implicit declarations and types which needs to be synthesised. ex: \Delta S needs to add and typecheck S~' and its parts)
ADD: while checking for consistency of the table created take into account multi-layered hiding/renaming (see complexHidingRenaming.tex for example). There is some repetition here in respect to bindins(Definition) functionality, but that's right in the sense that one code is auditing the other.

ADD: some cmd-line parameteres for the execution of DefTblService to enable more ellaborate testing

ADD: test cases on definition table services - it was disabled

TODO: handling of ZEves InstantiationLists in RenameExpr: this is tricky because the RHS is an expression to "rename"...