UPDATE: Updated reporting of parent cycles during parsing. Now the LatexMarkupParser also reports the warnings about cycles in parents. This was necessary, because if LatexMarkupParser gets to call the parents first, when the time comes to Parser and resetInfoTables(), the parents have already been parsed and cycles are no longer located.
Implementing this was a bit tricky, because at LatexMarkupParser we do not have Parent objects, and therefore their locations. We only have String parents. I have implemented a dynamically-resolving CyclicParentError, which allows deferring resolution of locations until the error is used (i.e. displayed in the UI). Then the error queries the section manager to locate the parsed specification and find out correct position for the given parent.

UPDATE: Refactor - moved reporting of parent cycles during parsing to static methods in CyclicParentError. This allowed to minimise impact onto the Parsers to some extent.