JAXBContext initialised as a static variable.
By initialising the JAXBContext statically for different JAXB Writers/Readers, we save a lot of processing time when the serialisation is used frequently. The JAXBContext depends on the context path and the class loader, which do not change. Furthermore, the context is thread safe. Therefore, it is ok to have it static, and avoid creating it for every serialisation (can take 0.5-1 sec).