ThreadLocal CyclicParseManager - workaround for bug #3511825.
Issues appear when the section manager is cloned for use in a multi-threaded environment. Since the CyclicParseManager is a singleton within a section manager, the same object is used after cloning due to shallow cloning. When several parse activities are performed in multiple threads, it fails, because it is designed to be used within a single parse.
As a workaround, made the CyclicParseManager ThreadLocal, so a different one will be used for other threads.

TODO: Review the singleton approach for CyclicParseManager.