Supporting toolkit file location in multiple JAR files.
When CZT core is package in one JAR (czt.jar), the toolkits are all collected together during the build. In a modular approach (separate JARs), this does not work, because the toolkits need to be resolved for different classes.

Implemented a workaround at the moment, where all providers are registered in /toolkit.loc file as classes from different packages. Then we use Class.getResource() for each provider to resolve them. This allows to locate toolkits in different JARs and still works for a single JAR build.