split Corejava into sub projects for each dialect;
After updating the GnAST with parameterised execution, split Corejava into per-dialect module structure. This allows clearer identification of dependencies as well as deployment of specific dialects only. Users can now make the dependencies smaller, e.g. if they are only interested in core Z and not interested in Object-Z.

After this change also dropped different version for Corejava (was 2.1.5 vs 1.6 for all of CZT). As Corejava is not deployed on its own anymore, reusing the main CZT version number for the submodules.

Note that during the split, all files are accounted for. They have only been moved, and new POM files have been added for submodules. This shows that the split was complete.