basic support for incremental build in GnAST;
Using Plexus BuildContext API to add incremental build support. This checks which files have changed and launches the build accordingly (tries not to regenerate unchanged files). When writing the file, also use the BuildContext so that the files are refreshed well.
By default, a non-incremental DefaultBuildContext is used which just forwards all operations to file system. So for standalone, the build works as before. When used within Eclipse, the incremental BuildContext is injected that allows only regenerate the affected files.