moved specreader to attic;
Specreader is not used anymore in any of the projects - decommissioned SpecSource as the last usage. Some tests are failing and are held together with a workaround (see bug "Specreader failing on Windows with CR+LF line endings"). The functionality is somewhat covered already by core CZT modules, e.g. support for "use before declaration", etc.