moved developer tools to /dev subdirectory.
The following projects have been moved: `gnast`, `gnast-maven-plugin`, `java-cup-runtime`, `java-cup`, `cup-maven-plugin`, and `parsergen-maven-plugin`.

The projects are used to build CZT, and thus are not part of CZT - e.g. the CZT users would not need them. Projects have been updated to be generic and configurable, so that they are more-or-less independent of CZT (e.g. can be configured via POM options). Therefore they should not be released every time with CZT releases, and should follow a different release cycle. Also, this would allow the CZT build be dependent on release versions of these DEV plugins/libraries.