Calculate APFD for no prioritisation case
The 'plot_apfd.py' script plotted and calculated APFD scores for the
implemented TCP systems. It did not calculate the APFD score when no
prioritisation was used.

The file and script added in this commit calculate the APFD score and
show a plot when no prioritistaion is used. That is, the default test
execution order when running `mvn clean install`.